fix: keep compatibility warning across app routes
